+++ cms/djangoapps/contentstore/errors.py
@@ -5,8 +5,6 @@ from django.utils.translation import ugettext as _
 
 COURSE_ALREADY_EXIST = _('Aborting import because a course with this id: {} already exists.')
 COURSE_PERMISSION_DENIED = _('Permission denied. You do not have write access to this course.')
-ERROR_WHILE_READING = _('Error while reading {}. Check file for XML errors.')
-FAILED_TO_IMPORT_MODULE = _('Failed to import module: {} at location: {}')
 FILE_MISSING = _('Could not find the {0} file in the package.')
 FILE_NOT_FOUND = _('Uploaded Tar file not found. Try again.')
 INVALID_FILE_TYPE = _('We only support uploading a .tar.gz file.')
